+++ wire/lib/src/hive/steps/activate.rs
@@ -61,7 +61,7 @@ async fn set_profile(
     let _ = child
         .wait_till_success()
         .await
-        .map_err(HiveLibError::DetachedError)?;
+        .map_err(HiveLibError::CommandError)?;
 
     info!("Set system profile");
 
@@ -143,7 +143,7 @@ impl ExecuteStep for SwitchToConfiguration {
                 let _ = reboot
                     .wait_till_success()
                     .await
-                    .map_err(HiveLibError::DetachedError)?;
+                    .map_err(HiveLibError::CommandError)?;
 
                 info!("Rebooted {name}, waiting to reconnect...", name = ctx.name);